+++ gcp_variant_transforms/pipeline_common.py
@@ -22,7 +22,6 @@ from typing import List  # pylint: disable=unused-import
 import argparse
 import enum
 import os
-import time
 import uuid
 from datetime import datetime
 
@@ -32,15 +31,12 @@ from apache_beam.io import filesystem
 from apache_beam.io import filesystems
 from apache_beam.options import pipeline_options
 from apache_beam.runners.direct import direct_runner
-from google.cloud import bigquery
 
 from gcp_variant_transforms.beam_io import bgzf_io
 from gcp_variant_transforms.beam_io import vcf_estimate_io
 from gcp_variant_transforms.beam_io import vcf_header_io
 from gcp_variant_transforms.beam_io import vcf_parser
 from gcp_variant_transforms.beam_io import vcfio
-from gcp_variant_transforms.libs import bigquery_util
-from gcp_variant_transforms.libs import sample_info_table_schema_generator
 from gcp_variant_transforms.transforms import fusion_break
 from gcp_variant_transforms.transforms import merge_headers
 
@@ -51,18 +47,6 @@ _LARGE_DATA_THRESHOLD = 50000
 
 _DATAFLOW_RUNNER_ARG_VALUE = 'DataflowRunner'
 SampleNameEncoding = vcf_parser.SampleNameEncoding
-
-_BQ_CREATE_PARTITIONED_TABLE_COMMAND = (
-    'bq mk --table '
-    '--range_partitioning=start_position,0,{TOTAL_BASE_PAIRS},{PARTITION_SIZE} '
-    '--clustering_fields=start_position,end_position '
-    '{FULL_TABLE_ID} {SCHEMA_FILE_PATH}')
-_BQ_DELETE_TABLE_COMMAND = 'bq rm -f -t {FULL_TABLE_ID}'
-_GCS_DELETE_FILES_COMMAND = 'gsutil -m rm -f -R {ROOT_PATH}'
-_BQ_LOAD_JOB_NUM_RETRIES = 5
-_MAX_NUM_CONCURRENT_BQ_LOAD_JOBS = 5
-_PARTITIONING_FIELD = 'start_position'
-_CLUSTERING_FIELDS = ['start_position', 'end_position']
 
 
 class PipelineModes(enum.Enum):
@@ -346,114 +330,3 @@ def generate_unique_name(job_name):
                    datetime.now().strftime('%Y%m%d-%H%M%S'),
                    str(uuid.uuid4())])
 
-
-def create_output_table(full_table_id, total_base_pairs, schema_file_path):
-  # type: (str, int, str) -> None
-  """Creates an integer range partitioned table using `bq mk table...` command.
-
-  Since beam.io.BigQuerySink is unable to create an integer range partition
-  we use `bq mk table...` to achieve this goal. Note that this command runs on
-  the worker that monitors the Dataflow job.
-
-  Args:
-    full_table_id: for example: projet:dataset.table_base_name__chr1
-    total_base_pairs: the maximum expected value of `start_position` column.
-    schema_file_path: a json file that contains the schema of the table.
-  """
-  (partition_size, total_base_pairs_enlarged) = (
-      bigquery_util.calculate_optimal_partition_size(total_base_pairs))
-  bq_command = _BQ_CREATE_PARTITIONED_TABLE_COMMAND.format(
-      TOTAL_BASE_PAIRS=total_base_pairs_enlarged,
-      PARTITION_SIZE=partition_size,
-      FULL_TABLE_ID=full_table_id,
-      SCHEMA_FILE_PATH=schema_file_path)
-  result = os.system(bq_command)
-  if result != 0:
-    raise ValueError(
-        'Failed to create a BigQuery table using "{}" command.'.format(
-            bq_command))
-
-
-def delete_table(full_table_id):
-  bq_command = _BQ_DELETE_TABLE_COMMAND.format(FULL_TABLE_ID=full_table_id)
-  return os.system(bq_command)
-
-
-def delete_gcs_files(root_path):
-  gcs_command = _GCS_DELETE_FILES_COMMAND.format(ROOT_PATH=root_path)
-  return os.system(gcs_command)
-
-
-class LoadAvro(object):
-  def __init__(self,
-               avro_root_path,  # type: str
-               table_base_name,  # type: str
-               suffixes,  # type: List[str]
-               total_base_pairs  # type: List[int]
-              ):
-    assert len(suffixes) == len(total_base_pairs)
-
-    self._avro_root_path = avro_root_path
-    self._table_base_name = table_base_name.replace(':', '.')
-    self._suffixes = suffixes
-    self._total_base_pairs = total_base_pairs
-
-    self._num_load_jobs_retries = 0
-    self._suffixes_to_load_jobs = {}  # type: Dict[str, bigquery.job.LoadJob]
-    self._remaining_load_jobs = self._suffixes
-
-    self._client = bigquery.Client()
-
-  def start_loading(self):
-    # We run _MAX_NUM_CONCURRENT_BQ_LOAD_JOBS load jobs in parallel.
-    for _ in range(_MAX_NUM_CONCURRENT_BQ_LOAD_JOBS):
-      self._start_one_load_job(self._remaining_load_jobs.pop())
-
-    self._monitor_load_jobs()
-
-  def _start_one_load_job(self, suffix):
-    job_config = bigquery.LoadJobConfig(
-        source_format=bigquery.SourceFormat.AVRO)
-    uri = self._avro_root_path + suffix + '-*'
-    table_id = sample_info_table_schema_generator.compose_table_name(
-        self._table_base_name, suffix)
-    load_job = self._client.load_table_from_uri(
-        uri, table_id, job_config=job_config)
-    self._suffixes_to_load_jobs.update({suffix: load_job})
-
-  def _cancel_all_running_load_jobs(self):
-    for _, load_job in self._suffixes_to_load_jobs.items():
-      load_job.cancel()
-
-  def _handle_failed_load_job(self, suffix, load_job):
-    if self._num_load_jobs_retries < _BQ_LOAD_JOB_NUM_RETRIES:
-      self._num_load_jobs_retries += 1
-      # Retry the failed job after 5 minutes wait.
-      time.sleep(300)
-      self._start_one_load_job(suffix)
-    else:
-      # Jobs have failed more than _BQ_LOAD_JOB_NUM_RETRIES, cancel all jobs.
-      self._cancel_all_running_load_jobs()
-      table_id = sample_info_table_schema_generator.compose_table_name(
-          self._table_base_name, suffix)
-      raise ValueError(
-          'Failed to load AVRO to BigQuery table {} \n state: {} \n '
-          'job_id: {} \n errors: {}.'.format(table_id, load_job.state,
-                                             load_job.path,
-                                             '\n'.join(load_job.errors)))
-  def _monitor_load_jobs(self):
-    # Waits until current jobs are done and then add remaining jobs one by one.
-    while self._suffixes_to_load_jobs:
-      time.sleep(60)
-      processed_suffixes = self._suffixes_to_load_jobs.keys()
-      for suffix in processed_suffixes:
-        load_job = self._suffixes_to_load_jobs.get(suffix)
-        if load_job.done():
-          del self._suffixes_to_load_jobs[suffix]
-          state = load_job.state
-          if state != 'DONE':
-            self._handle_failed_load_job(suffix, load_job)
-          else:
-            if self._remaining_load_jobs:
-              next_suffix = self._remaining_load_jobs.pop()
-              self._start_one_load_job(next_suffix)
